Media-grid now has minimum widths - heights aren't always the same (are close)
Instead of squashing all items onto one line, even if that means they're very small, I specified a min-width of 25% (after padding, this means I should be able to have 3 items max per line).

This means that sometimes to achieve this I can't have them be the same height -if we have images with very opposite aspect ratios - e.g. 2x3 vs 3x2 - then they need a 20/80 not 25/75 split, to be the same height. That's okay, this is good enough for now.